publishing of prosperity values is now correct. TODO: the prosperity values probably are not very informative, maybe sliding window for n last steps

+++ sarsa/src/main/java/org/hanns/rl/discrete/observer/stats/combined/ForgettingCoverageChangeReward.java
@@ -64,7 +64,9 @@ public class ForgettingCoverageChangeReward extends AbsProsperityObserver{
 
 	@Override
 	public float getProsperity() {
-		return (cover.getProsperity()+rew.getProsperity()+ch.getProsperity())/3;
+		return (float)(((double)cover.getProsperity()
+				+(double)rew.getProsperity()
+				+(double)ch.getProsperity())/3.0);
 	}
 
 	@Override
